[FLINK-21945][checkpoint] Ensure that each operator is stored in the checkpoint, possibly with empty state.
Unaligned checkpoint recovery needs to know the old parallelism pre-rescaling of both upstream and downstream side. Currently, the state is omitted when an operator has no state at all. However, if the downstream operator is omitted, then it's impossible to calculate the channel mapping for upstream and vice versa.

The change adds a couple of bytes to the checkpoint per operator (not suboperator). Thus, except for long stateless pipelines, the change would be barely noticeable.