[FLINK-24038] Disable leadership tests when using multiple component leadership ha services
This commit disables the leadership tests done by the KubernetesCheckpointIDCounter and the KubernetesStateHandleStore
when using the multiple component leadership ha services. This is necessary because when using this ha service, then the
config maps used by KubernetesCheckpointIDCounter and KubernetesStateHandlStore have not LeaderElector assigned. Moreover,
this commit explicitly creates the required config maps which have been created before by the LeaderElector.