[FLINK-25855] Allow JobMaster to accept excess slots when restarting the job
This commit allows the JobMaster to accept excess slots when the job is restarting.
The way it works is that the DefaultScheduler tells the DefaultSlotPoolBridge when
the job is restarting. If during this time new slots should be offered, then the
DefaultSLotPoolBridge calls DeclarativeSlotPool.registerSlots which accepts all
slot offers. Once the job leaves the RESTARTING state, the DefaultSlotPoolBridge
will call DeclarativeSlotPool.offerSlots which only accepts those slots that are
currently required.

[FLINK-25855] Add DeclarativeSlotPool.registerSlots

This closes #18542.