[FLINK-25982] Support idleness with watermark alignment
When a SourceOperator becomes idle it reports MAX_WATERMARK to the
source coordinator, allowing other participating tasks/sources to
progress their watermark. As soon as the operator becomes active it
sends an actual watermark. Either a new, updated or the old one if it
became active without increasing the watermark.